01_ai_agents_first/appendix/_deployment/02_advanced_config/pyproject.toml
[project]
name = "chainlit-advanced"
version = "0.1.0"
description = "Advanced Chainlit deployment with custom features"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.4.0",
    "google-generativeai>=0.3.2",
    "python-dotenv>=1.0.1",
    "pyjwt>=2.8.0",
    "python-jose>=3.3.0",
    "httpx>=0.26.0",
    "typing-extensions>=4.5.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 